requirements.txt
mysql-connector-python
pandas
sqlalchemy
PyMySQL
